+++ app/controllers/public/bookings_controller.rb
@@ -11,12 +11,15 @@ module Public
 
     def new
       @booking = preparation_service.prepare_new(create_params)
+      BookingQuestionResponse.prepare_booking(@booking)
 
       respond_with :public, @booking
     end
 
     def edit
       @booking.committed_request ||= @booking.agent_booking&.committed_request if @booking.agent_booking.present?
+      BookingQuestionResponse.prepare_booking(@booking)
+
       respond_with :public, @booking
     end
 
@@ -34,10 +37,8 @@ module Public
     end
 
     def update
-      if @booking.editable?
-        @booking.assign_attributes(update_params)
-        @booking.save(context: :public_update)
-      end
+      @booking.assign_attributes(update_params) if @booking.editable?
+      @booking.save(context: :public_update)
       call_booking_action
       Booking::Log.log(@booking, trigger: :tenant, action: booking_action, user: current_user)
       respond_with :public, @booking, location: edit_public_booking_path(@booking.token)
@@ -67,6 +68,11 @@ module Public
     rescue BookingActions::Base::NotAllowed
       @booking.errors.add(:base, :action_not_allowed)
     end
+
+    # def process_booking_questions
+    #   sanitized_params = params.dig(:booking, :booking_questions)
+    #   @booking.booking_questions = Manage::BookingQuestionParams.sanitize_booking_params(@booking, sanitized_params)
+    # end
 
     def booking_action
       BookingActions::Public.all[params[:booking_action]]