+++ app/serializers/manage/organisation_serializer.rb
@@ -5,9 +5,11 @@ module Manage
     association :booking_categories, blueprint: BookingCategorySerializer
     association :rich_text_templates, blueprint: RichTextTemplateSerializer
     association :homes, blueprint: HomeSerializer
-    association :tenants, blueprint: TenantSerializer
+    # association :tenants, blueprint: TenantSerializer
     association :designated_documents, blueprint: DesignatedDocumentSerializer
     association :tarifs, blueprint: TarifSerializer
+    association :booking_categories, blueprint: BookingCategorySerializer
+    association :booking_questions, blueprint: Public::BookingQuestionSerializer
 
     fields :qr_iban, :esr_beneficiary_account, :iban, :mail_from
     field :settings do |organisation|