+++ app/serializers/public/booking_question_response_serializer.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+module Public
+  class BookingQuestionResponseSerializer < ApplicationSerializer
+    association :booking_question, blueprint: Public::BookingQuestionSerializer
+
+    fields :value, :booking_question_id
+  end
+end