+++ spec/models/booking_question_response_spec.rb
@@ -1,0 +1,26 @@
+# frozen_string_literal: true
+
+# == Schema Information
+#
+# Table name: booking_question_responses
+#
+#  id                  :bigint           not null, primary key
+#  value               :jsonb
+#  created_at          :datetime         not null
+#  updated_at          :datetime         not null
+#  booking_id          :uuid             not null
+#  booking_question_id :bigint           not null
+#
+# Indexes
+#
+#  index_booking_question_responses_on_booking_question_id  (booking_question_id)
+#
+# Foreign Keys
+#
+#  fk_rails_...  (booking_question_id => booking_questions.id)
+#
+require 'rails_helper'
+
+RSpec.describe BookingQuestionResponse, type: :model do
+  pending "add some examples to (or delete) #{__FILE__}"
+end