+++ spec/models/payment_infos/qr_bill_spec.rb
@@ -47,4 +47,11 @@ RSpec.describe PaymentInfos::QrBill, type: :model do
       it { is_expected.to eq('RF 37123 45678 91011 1213') }
     end
   end
+
+  describe 'prepare_address_lines' do
+    let(:address_lines) { %w[Line1 Line2 Line3 Line4] }
+    subject { qr_bill.send(:prepare_address_lines, address_lines) }
+
+    it { is_expected.to eq(%w[Line1 Line2 Line3]) }
+  end
 end